Remove dolfin_adjoint_common and compat abstraction
