--- dolfin_adjoint_common/blocks/__init__.py
@@ -1,8 +1,0 @@
-# flake8: noqa
-
-from .assembly import AssembleBlock
-from .solving import GenericSolveBlock
-from .function import FunctionAssignBlock
-from .dirichlet_bc import DirichletBCBlock
-from .constant import ConstantAssignBlock
-
