+++ fenics_adjoint/blocks/__init__.py
@@ -1,11 +1,13 @@
 # flake8: noqa
 
-from .common import *
 from .solving import SolveLinearSystemBlock, SolveVarFormBlock
 from .projection import ProjectBlock
 from .variational_solver import LinearVariationalSolveBlock, NonlinearVariationalSolveBlock
 from .krylov_solver import KrylovSolveBlock, KrylovSolveBlockHelper
 from .lu_solver import LUSolveBlock, LUSolveBlockHelper
 from .function_assigner import FunctionAssignerBlock
-from fenics_adjoint.blocks.function import FunctionEvalBlock, FunctionSplitBlock, FunctionMergeBlock
+from .function import FunctionEvalBlock, FunctionSplitBlock, FunctionMergeBlock
 from .petsc_krylov_solver import PETScKrylovSolveBlock, PETScKrylovSolveBlockHelper
+from .constant import ConstantAssignBlock
+from .dirichlet_bc import DirichletBCBlock
+from .assembly import AssembleBlock
