+++ fenics_adjoint/blocks/krylov_solver.py
@@ -1,8 +1,10 @@
 import dolfin
-from . import SolveLinearSystemBlock
-from dolfin_adjoint_common import compat
 
-compat = compat.compat(dolfin)
+from fenics_adjoint.utils import function_from_vector
+
+
+from . import SolveLinearSystemBlock
+from .assembly import assemble_adjoint_value
 
 
 class KrylovSolveBlockHelper(object):
@@ -70,12 +72,12 @@ class KrylovSolveBlock(SolveLinearSystemBlock):
                 else:
                     solver.set_operator(A)
             else:
-                A = compat.assemble_adjoint_value(dFdu_adj_form)
+                A = assemble_adjoint_value(dFdu_adj_form)
                 [bc.apply(A) for bc in bcs]
 
                 if self.pc_operator is not None:
                     P = self._replace_form(self.pc_operator)
-                    P = compat.assemble_adjoint_value(P)
+                    P = assemble_adjoint_value(P)
                     [bc.apply(P) for bc in bcs]
                     solver.set_operators(A, P)
                 else:
@@ -91,7 +93,7 @@ class KrylovSolveBlock(SolveLinearSystemBlock):
 
         adj_sol_bdy = None
         if compute_bdy:
-            adj_sol_bdy = compat.function_from_vector(self.function_space, dJdu_copy - compat.assemble_adjoint_value(
+            adj_sol_bdy = function_from_vector(self.function_space, dJdu_copy - assemble_adjoint_value(
                 dolfin.action(dFdu_adj_form, adj_sol)))
 
         return adj_sol, adj_sol_bdy
@@ -109,11 +111,11 @@ class KrylovSolveBlock(SolveLinearSystemBlock):
                 else:
                     solver.set_operator(A)
             else:
-                A = compat.assemble_adjoint_value(lhs)
+                A = assemble_adjoint_value(lhs)
                 [bc.apply(A) for bc in bcs]
                 if self.pc_operator is not None:
                     P = self._replace_form(self.pc_operator)
-                    P = compat.assemble_adjoint_value(P)
+                    P = assemble_adjoint_value(P)
                     [bc.apply(P) for bc in bcs]
                     solver.set_operators(A, P)
                 else:
@@ -125,7 +127,7 @@ class KrylovSolveBlock(SolveLinearSystemBlock):
             b = dolfin.Function(self.function_space).vector()
             system_assembler.assemble(b)
         else:
-            b = compat.assemble_adjoint_value(rhs)
+            b = assemble_adjoint_value(rhs)
             [bc.apply(b) for bc in bcs]
 
         solver.parameters.update(self.krylov_solver_parameters)
