+++ fenics_adjoint/blocks/petsc_krylov_solver.py
@@ -1,9 +1,10 @@
 import dolfin
-from . import SolveLinearSystemBlock
-from fenics_adjoint.types import as_backend_type
-from dolfin_adjoint_common import compat
 
-compat = compat.compat(dolfin)
+from fenics_adjoint.utils import function_from_vector, as_backend_type
+
+
+from . import SolveLinearSystemBlock
+from .assembly import assemble_adjoint_value
 
 
 class PETScKrylovSolveBlockHelper(object):
@@ -72,7 +73,7 @@ class PETScKrylovSolveBlock(SolveLinearSystemBlock):
                 else:
                     solver.set_operator(A)
             else:
-                A = compat.assemble_adjoint_value(dFdu_adj_form)
+                A = assemble_adjoint_value(dFdu_adj_form)
                 [bc.apply(A) for bc in bcs]
 
                 if self._ad_nullspace is not None:
@@ -80,7 +81,7 @@ class PETScKrylovSolveBlock(SolveLinearSystemBlock):
 
                 if self.pc_operator is not None:
                     P = self._replace_form(self.pc_operator)
-                    P = compat.assemble_adjoint_value(P)
+                    P = assemble_adjoint_value(P)
                     [bc.apply(P) for bc in bcs]
                     solver.set_operators(A, P)
                 else:
@@ -100,7 +101,7 @@ class PETScKrylovSolveBlock(SolveLinearSystemBlock):
 
         adj_sol_bdy = None
         if compute_bdy:
-            adj_sol_bdy = compat.function_from_vector(self.function_space, dJdu_copy - compat.assemble_adjoint_value(
+            adj_sol_bdy = function_from_vector(self.function_space, dJdu_copy - assemble_adjoint_value(
                 dolfin.action(dFdu_adj_form, adj_sol)))
 
         return adj_sol, adj_sol_bdy
@@ -124,14 +125,14 @@ class PETScKrylovSolveBlock(SolveLinearSystemBlock):
                 else:
                     solver.set_operator(A)
             else:
-                A = compat.assemble_adjoint_value(lhs)
+                A = assemble_adjoint_value(lhs)
                 [bc.apply(A) for bc in bcs]
                 if self._ad_nullspace is not None:
                     as_backend_type(A).set_nullspace(self._ad_nullspace)
 
                 if self.pc_operator is not None:
                     P = self._replace_form(self.pc_operator)
-                    P = compat.assemble_adjoint_value(P)
+                    P = assemble_adjoint_value(P)
                     [bc.apply(P) for bc in bcs]
                     solver.set_operators(A, P)
                 else:
@@ -143,7 +144,7 @@ class PETScKrylovSolveBlock(SolveLinearSystemBlock):
             b = dolfin.Function(self.function_space).vector()
             system_assembler.assemble(b)
         else:
-            b = compat.assemble_adjoint_value(rhs)
+            b = assemble_adjoint_value(rhs)
             [bc.apply(b) for bc in bcs]
 
         if self._ad_nullspace is not None:
