--- fenics_adjoint/compat.py
@@ -1,7 +1,0 @@
-from dolfin_adjoint_common.compat import compat
-import dolfin
-
-
-class Backend:
-    backend = dolfin
-    compat = compat(dolfin)
