+++ fenics_adjoint/petsc_krylov_solver.py
@@ -1,11 +1,10 @@
 import dolfin
 
 from pyadjoint.tape import annotate_tape, get_working_tape
-from dolfin_adjoint_common import compat
 
-from .blocks import PETScKrylovSolveBlock, PETScKrylovSolveBlockHelper
 
-compat = compat.compat(dolfin)
+from fenics_adjoint.blocks import PETScKrylovSolveBlock, PETScKrylovSolveBlockHelper
+from fenics_adjoint.utils import MatrixTypes
 
 
 class PETScKrylovSolver(dolfin.PETScKrylovSolver):
@@ -18,10 +17,10 @@ class PETScKrylovSolver(dolfin.PETScKrylovSolver):
         preconditioner = kwargs.pop("preconditioner", "default")
 
         next_arg_idx = 0
-        if len(args) > 0 and isinstance(args[0], compat.MatrixType):
+        if len(args) > 0 and isinstance(args[0], MatrixTypes):
             A = args[0]
             next_arg_idx = 1
-        elif len(args) > 1 and isinstance(args[1], compat.MatrixType):
+        elif len(args) > 1 and isinstance(args[1], MatrixTypes):
             A = args[1]
             next_arg_idx = 2
