+++ fenics_adjoint/types/genericvector.py
@@ -1,13 +1,12 @@
 import dolfin
-from dolfin_adjoint_common import compat
-compat = compat.compat(dolfin)
+from fenics_adjoint.utils import gather
 
 __all__ = []
 
 
 @staticmethod
 def _ad_to_list(self):
-    return compat.gather(self).tolist()
+    return gather(self).tolist()
 
 
 dolfin.GenericVector._ad_to_list = _ad_to_list
