+++ setup.py
@@ -19,12 +19,9 @@ setup(name='dolfin_adjoint',
       packages=['fenics_adjoint',
                 'fenics_adjoint.types',
                 'fenics_adjoint.blocks',
-                'dolfin_adjoint',
-                'dolfin_adjoint_common',
-                'dolfin_adjoint_common.blocks'],
+                'dolfin_adjoint'],
       package_dir={'fenics_adjoint': 'fenics_adjoint',
-                   'dolfin_adjoint': 'dolfin_adjoint',
-                   'dolfin_adjoint_common': 'dolfin_adjoint_common'},
+                   'dolfin_adjoint': 'dolfin_adjoint'},
       install_requires=['scipy>=1.0', 'pyadjoint@git+https://github.com/dolfin-adjoint/pyadjoint'],
       extras_require=extras
       )
