+++ tests/fenics_adjoint/test_adjoint_state_cb.py
@@ -5,6 +5,7 @@ from fenics import *
 from fenics_adjoint import *
 
 
+
 def test_time_dependent():
     mesh = UnitSquareMesh(5, 5)
     V = FunctionSpace(mesh, "CG", 1)
@@ -19,8 +20,8 @@ def test_time_dependent():
     dt = 0.1
     T = 1.0
 
-    a = inner(u, v)*dx + dt * inner(grad(u), grad(v))*dx
-    L = inner(U_prev, v)*dx
+    a = inner(u, v) * dx + dt * inner(grad(u), grad(v)) * dx
+    L = inner(U_prev, v) * dx
 
     adj_states = []
     adj_cb = lambda adj_sol: adj_states.append(adj_sol)
@@ -31,12 +32,11 @@ def test_time_dependent():
         t += dt
         U_prev.assign(U)
 
-    J = assemble(inner(U, U)*dx)
+    J = assemble(inner(U, U) * dx)
     Jhat = ReducedFunctional(J, control)
     dJdm = Jhat.derivative()
 
-    dFdm = -inner(u, v)*dx
+    dFdm = -inner(u, v) * dx
     adj_int = adj_states[-1]
     dJdm_manual = -assemble(action(dFdm, adj_int))
     assert abs(dJdm.vector().get_local() - dJdm_manual.get_local()).sum() <= 0
-
